==> 99.run_all_preceding_Rmd_files.Rmd <==
---
title: "A good title to conceive"
author: "Author: Old Vegetable Bird"
date: "`r format(Sys.time(), '%F')`"
output:
    rmarkdown::html_document:
        theme: readable
        highlight: textmate
        df_print: paged
---

```{r message=FALSE, warning=FALSE}
rm(list = ls())
ptm <- proc.time()
library("magrittr")
# proc.time() - ptm
```

# Refresh run

```{r}
output_files <- here::here("output_files") %T>%
    dir.create(., showWarnings = F, recursive = T)
unlink(file.path(output_files, "*"), recursive = T, force = T)
```

# List Rmd files to run, in other words, render.

```{r}
serial_scripts <- fs::dir_ls(here::here(), glob = "*.Rmd") %>%
    .[-length(.)]
serial_scripts
```

# Run all the Rmd files, namely render them.

```{r message=FALSE, warning=FALSE}
for (i in serial_scripts) {
    rmarkdown::render(i, quiet = TRUE)
}
```

# Clean up rendered html files.

```{r}
fs::dir_ls(here::here(), glob = "*.html") %>%
    fs::file_delete(.)
```

```{r}
proc.time() - ptm
```
